qwen-edit-lab/requirements.txt
# requirements.txt
torch
diffusers
transformers
accelerate
Pillow
invisible-watermark
fastapi
